Refactor and enhance DXF serialization support
Refactored `CommonPolylineConverter` to remove `PolyfaceMeshConverter` and added its implementation in a new file with JSON deserialization support. Updated `ImageDefinition` and `VisualStyle` classes to include additional properties, XML documentation, and DXF serialization attributes. Improved code structure, readability, and maintainability.
